split in vtable case to stablize ci running
